feat: enhance container dependency injection, improve relationship eager loading, and optimize route resolution
- Container: Add lazy singleton support, enhance service provider binding, improve error handling
- Eloquent: Refactor relationship eager loading for better performance and N+1 query prevention
- Relationships: Add HasOne, HasMany, BelongsTo relationship caching and query optimization
- Schema: Improve schema introspection and column type detection
- Routing: Optimize route resolution performance with better caching and parameter handling
- Casts: Simplify primitive type casting logic for better maintainability

+++ cara/eloquent/schema/Schema.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from cara.facades import DB
 
 from .Blueprint import Blueprint
@@ -35,7 +37,7 @@ class Schema:
         "uuid": str,
         "binary": bytes,
         "boolean": bool,
-        "decimal": float,
+        "decimal": Decimal,
         "double": float,
         "enum": str,
         "text": str,
